+++ src/cirq_iqm/iqm_remote.py
@@ -37,11 +37,7 @@ def serialize_circuit(circuit: cirq.Circuit) -> CircuitDTO:
     Returns:
         data transfer object representing the circuit
     """
-    instructions = []
-    for moment in circuit.moments:
-        for operation in moment.operations:
-            instructions.append(map_operation(operation))
-
+    instructions = list(map(map_operation, circuit.all_operations()))
     return CircuitDTO(
         name='Serialized from Cirq',
         instructions=instructions,
@@ -125,10 +121,9 @@ class IQMSampler(cirq.work.Sampler):
         # check that the circuit connectivity fits in the device connectivity
         # apply qubit_mapping
         qubit_map = {cirq.NamedQubit(k): cirq.NamedQubit(v) for k, v in self._qubit_mapping.items()}
-        qubit_map = program.transform_qubits(temp)
-        for m in mapped.moments:
-            for op in m.operations:
-                self._device.check_qubit_connectivity(op)
+        mapped = program.transform_qubits(qubit_map)
+        for op in mapped.all_operations():
+            self._device.check_qubit_connectivity(op)
 
         results = [self._send_circuit(program, repetitions=repetitions)]
         return results