+++ CHANGELOG.rst
@@ -17,8 +17,9 @@ Features
   * ``IQMSampler`` checks that the circuit respects the device connectivity.
 
 * Device qubit handling is simplified. `#34 <https://github.com/iqm-finland/cirq-on-iqm/pull/34>`_
+
   * ``IQMSampler`` can generate a trivial qubit mapping automatically.
-  * The class ``IQMQubit`` removed.
+  * The class ``IQMQubit`` was removed.
 
 Bugfixes
 --------